move db utils
